Setup for internal sso(openidc)
This patch prepares Keycloak realm and client for ovirt engine internal
sso needs.
This includes:
- creating dedicated Keycloak realm, client and necessary scopes.
- creating Keycloak internal admin user that mimics already existing admin@internal
- preparing ovirt engine extension configuration files
- preparing httpd openidc configuration

Very important assumption is that keycloak based internal sso is enabled
only for *new* and *RPM* based deployment of ovirt-engine. That means that
authentication for development mode stays as it is as well as for
upgraded installations.

Future steps (outside of the scope of this patch) includes:
- replacing jdbc based internal sso for non-dev environments
- preparing standalone tool for jdbc users migration for upgrading
  existing installation.

Signed-off-by: Artur Socha <asocha@redhat.com>